(#10) WallCollision: split into server and client
