+++ Source/Client/General/ClientPhysicsState.cs
@@ -14,4 +14,10 @@ public class ClientPhysicsState : PhysicsState
     {
         return new ClientPlayerCollision(plr, pos, sv, radius, isplayer);
     }
+
+    protected override WallCollision CreateWallCollision(Linedef ld, Vector3D sv, float stepheight)
+    {
+        // Make new collision with updated parameters
+        return new ClientWallCollision(ld, pos, sv, radius, height, stepheight, isplayer);
+    }
 }
