+++ Source/Server/General/ServerPhysicsState.cs
@@ -11,5 +11,10 @@ public class ServerPhysicsState : PhysicsState
         return new ServerPlayerCollision(plr, pos, sv, radius, isplayer);
     }
 
+    protected override WallCollision CreateWallCollision(Linedef ld, Vector3D sv, float stepheight)
+    {
+        return new ServerWallCollision(ld, pos, sv, radius, height, stepheight, isplayer);
+    }
+
     protected override bool IsClientMode => false;
 }
